Move ndc-test functions into their own library.
